requirements.txt
aiogram==3.10.0
python-dotenv==1.0.0
aiohttp==3.9.0
nest-asyncio==1.6.0
boto3==1.34.0  
python-dateutil==2.8.2  
async-timeout==4.0.3
bcrypt